+++ CHANGELOG.md
@@ -16,9 +16,12 @@ The format is based on [Keep a Changelog].
 ### Removed
 
 ### Fixed
-- Unexpected errors with long window size `#46`
 
 ### Security
+
+## [0.1.12] - 2026-03-23
+### Fixed
+- Unexpected errors with long window size [`#46`](https://github.com/anatawa12/deflate64-rs/pull/46)
 
 ## [0.1.11] - 2026-02-26
 ### Added
@@ -80,7 +83,8 @@ The format is based on [Keep a Changelog].
 ### Added
 - Initial Deflate64 implementation
 
-[Unreleased]: https://github.com/anatawa12/deflate64-rs/compare/v0.1.11...HEAD
+[Unreleased]: https://github.com/anatawa12/deflate64-rs/compare/v0.1.12...HEAD
+[0.1.12]: https://github.com/anatawa12/deflate64-rs/compare/v0.1.11...v0.1.12
 [0.1.11]: https://github.com/anatawa12/deflate64-rs/compare/v0.1.10...v0.1.11
 [0.1.10]: https://github.com/anatawa12/deflate64-rs/compare/v0.1.9...v0.1.10
 [0.1.9]: https://github.com/anatawa12/deflate64-rs/compare/v0.1.8...v0.1.9